Fix type casting

+++ tests/global_data_usage/cvt_f64_i32/cvt_f64_i32_hands.cl
@@ -5,5 +5,5 @@ void float_test(__global double *out, int i)
 {
     int var0;
     var0 = gdata0[get_global_id(0)];
-    out[get_global_id(0)] = var0;
+    out[get_global_id(0)] = (double)var0;
 }
