+++ tests/global_data_usage/cvt_i32_f64/cvt_i32_f64_hands.cl
@@ -5,5 +5,5 @@ void float_test(__global int *out, int i)
 {
     double var0;
     var0 = gdata0[get_global_id(0)];
-    out[get_global_id(0)] = var0;
+    out[get_global_id(0)] = (int)var0;
 }
